code reloading on server side works
had to use trio; getting the async watch function and server to run concurrently and be ctrl-C-able is too much to ask of asyncio :joy:

for now user code runs in threadpool (fastapi's?). eventually we want processes, but it's too wild for the prototype.

user code that spans multiple files almost certainly won't work!!